BKO now moves competition to the next phase (RK/BK)
